+++ util/configuration.py
@@ -1,8 +1,8 @@
-from .mielib import custominput as ci
-from .extension import cleanString
-import base64
-import yaml
 import os
+import yaml
+
+from .mielib import custominput as ci
+from .extension import cleanString, decode, encode
 
 class File:
     __util_dir = os.path.dirname(__file__)
@@ -145,7 +145,7 @@ class Email:
     address = __data.get("address", "<your.email@gmail.com>")
     password = ""
     try:
-        password = base64.b64decode(__data.get("password")).decode('utf-8')
+        password = decode(__data.get("password"))
     except:
         password = "<your password>"
     server = __data.get("server", "smtp.gmail.com")
@@ -157,11 +157,11 @@ class Email:
         email_address = ci.email_input("What is the gmail address you would " \
             "like me to use to send you reports?", provider="gmail")
         password = ci.password_input("What is the password to the account you" \
-            " just entered?").encode('utf-8')
+            " just entered?")
         recipients = ci.email_input("What email address(es) would you like " \
             "to recieve the logs and reports?", multiples=True)
         cls.address = email_address
-        cls.password = base64.b64encode(password)
+        cls.password = password
         cls.recipients = recipients
         cls.update()
 
@@ -192,55 +192,95 @@ class Maintenance:
     backup_schedule = __backup.get("schedule", "0 3 * * *")
     backup_path = __backup.get("path", "~/MC_Backups")
     backup_number = __backup.get("number", 1)
+    backup_file_server = __backup.get("file_server", {})
     maintenance_running = __data.get('scheduled_running', False)
     update_schedule = __update.get("schedule", "0 3 * * 0")
     update_allow_major_update = __update.get("allow_major_update", False)
 
     @classmethod
     def build(cls):
-        print("Warning: A system restart is good practice to clear out any " \
-            "residual problems that might still be in RAM. Also, in order to " \
-            "run the commands file a server restart is required.")
-        restart_cron = ci.cron_date_input("restart")
+        # print("Warning: A system restart is good practice to clear out any " \
+        #     "residual problems that might still be in RAM. Also, in order to " \
+        #     "run the commands file a server restart is required.")
+        # restart_cron = ci.cron_date_input("restart")
 
-        print("Warning: It is good practice to backup your server so if any" \
-            "thing were to happen, you would be able to revert back to your " \
-            "previous backup.")
-        backup_cron = ci.cron_date_input("backup Minecraft")
-        backup_path = input("Where would you like your backups to be stored? ")
-        backup_limit = ci.int_input("How many backups would you like to be " \
-            "stored before removing old backups?")
+        # print("Warning: It is good practice to backup your server so if any" \
+        #     "thing were to happen, you would be able to revert back to your " \
+        #     "previous backup.")
+        # backup_cron = ci.cron_date_input("backup Minecraft")
+        # backup_path = input("Where would you like your backups to be stored? ")
+        # backup_limit = ci.int_input("How many backups would you like to be " \
+        #     "stored before removing old backups?")
+        file_server = cls.__build_external_storage()
+        print(file_server)
 
-        print("Warning: It is wise to check for updates on a regular basis so " \
-            "any bugs the developers might find and fix will be applied to " \
-            "your server. We can understand your concern for larger updates, " \
-            "so we will ask your permission on if you want us to do bigger " \
-            "updates automatically. If not, we will email you and alert you " \
-            "of any major updates.")
-        update_cron = ci.cron_date_input("check for updates")
-        major_updates = ci.bool_input("Would you like me to update to " \
-            "major releases?", default=False)
+        # print("Warning: It is wise to check for updates on a regular basis so " \
+        #     "any bugs the developers might find and fix will be applied to " \
+        #     "your server. We can understand your concern for larger updates, " \
+        #     "so we will ask your permission on if you want us to do bigger " \
+        #     "updates automatically. If not, we will email you and alert you " \
+        #     "of any major updates.")
+        # update_cron = ci.cron_date_input("check for updates")
+        # major_updates = ci.bool_input("Would you like me to update to " \
+        #     "major releases?", default=False)
             
-        print("Warning: I have ben preprogrammed with some useful maintenance " \
-            "scripts to help keep your server up and running smoothly. It is " \
-            "always good to run these scripts so your players experience as " \
-            "little server lag as possible.")
-        maintenance_cron = ci.cron_date_input("run maintenance scripts")
+        # print("Warning: I have ben preprogrammed with some useful maintenance " \
+        #     "scripts to help keep your server up and running smoothly. It is " \
+        #     "always good to run these scripts so your players experience as " \
+        #     "little server lag as possible.")
+        # maintenance_cron = ci.cron_date_input("run maintenance scripts")
 
-        cls.complete_shutdown = restart_cron
-        cls.schedule = maintenance_cron
-        cls.backup_schedule = backup_cron
-        cls.backup_path = backup_path
-        cls.backup_number = backup_limit
-        cls.update_schedule = update_cron
-        cls.update_allow_major_update = major_updates
+        # cls.complete_shutdown = restart_cron
+        # cls.schedule = maintenance_cron
+        cls.backup_file_server = file_server
+        # cls.backup_schedule = backup_cron
+        # cls.backup_path = backup_path
+        # cls.backup_number = backup_limit
+        # cls.update_schedule = update_cron
+        # cls.update_allow_major_update = major_updates
         cls.update()
 
     @classmethod
+    def __build_external_storage(cls) -> object:
+        '''
+        Requests user input regarding their external file server where backups
+        will be stored.
+
+        Returns: An object containing all of the needed values for configuration.
+        '''
+        should_setup = ci.bool_input('For extra data security, I can also ' \
+            'store your backups on an external file security. Would you like ' \
+            'to set that up now? (Please ensure that this machine has ' \
+            'connected to your file server before!)')
+        if should_setup:
+            host = ci.server_address_input('What is the host address of your ' \
+                'file server?')
+            print(f'Collecting ssh key for {host}...')
+            key = os.popen(f'ssh-keyscan {host} | grep "ssh-rsa"').read()
+            key = key.replace(f'{host} ssh-rsa ', '')
+            encoded_key = encode(key)
+            username = ci.string_input('What is the username for your file ' \
+                'server?')
+            password = ci.password_input('What is the password used to ' \
+                'connect to your file server?')
+            path = ci.string_input('Where would you like to store backups on ' \
+                'your file server?', r'^~?(?:\/.+){1,}', '~/backups')
+            return {
+                'domain' : host,
+                'key' : encoded_key,
+                'username' : username,
+                'password' : password,
+                'path' : path
+            }
+        return {}
+
+    @classmethod
     def update(cls):
+        cls.__backup['file_server'] = cls.backup_file_server
         cls.__backup["schedule"] = cls.backup_schedule
         cls.__backup["path"] = cls.backup_path
         cls.__backup["number"] = cls.backup_number
+        cls.__backup["file_server"] = cls.backup_file_server
         cls.__update["schedule"] = cls.update_schedule
         cls.__update["allow_major_update"] = cls.update_allow_major_update
         cls.__data["complete_shutdown"] = cls.complete_shutdown
@@ -259,6 +299,7 @@ class Maintenance:
         cls.backup_number = 1
         cls.update_schedule = "0 3 * * 0"
         cls.update_allow_major_update = False
+        cls.update()
 
 class Messaging:
     __data = File.data.get('Messaging', {})