+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- The user can now add configuration for uploading server backups to an external file server.
+  - This configuration exists under the Maintenance section of the config and can be updated by calling `python main.py -uc maintenance`.
+  - If the user has configured these settings and the project fails to upload a backup to the server for any reason, the project generates an email to send to the server admins.
 
 ### Changed